Adding Common Error Message:  "Task 1, We are unable to connect to the source site.  Please check the source"
I am submitting to the PG as well to update the error message to state "We are unable to connect to the target site."  which would make more sense since the issue can be with the target.

While they consider this I am adding the Common error because my customer faced this issue two times in the same week resulting in support cases.   Adding the error to this article hopefully will provide some guidance for other customers who encounter this.  Please contact me if additional information is required to add this to the document.