Update pre-provision-accounts.md
updating the code snippet. 
the previous code snippet will not work with Graph Cmdlets as the commands used have changed and require significant changes to the parameters, furthermore there are strong arguments on regards of security on regards of authentication with the previous script. (if they were to work) 

the previous code will fail to work with the following errors: 

Connect-MgGraph : Tenant id cannot be null. 
Connect-SPOService : One or more errors occurred.
At line:3 char:1
+ Connect-SPOService -Credential $Credential -Url ...
Get-MgUser : Authentication needed. Please call Connect-MgGraph.

also, the usage of credentials is discouraged (if not plainly disabled at this point) you should use MSAL/Oauth instead. and to do so you don't need to provide the "-Credential" 

the new script corrects this nuance and include the Tenant flag. also remove the variable for the URL from Contoso to a parameter so the User of the script can provide the value without changing the script. 

if need be, an example usage can be provided such as 


$SharepointURL = "https://contoso.sharepoint.com"
$tenantID = "xxxx.xxxx.xxx"
.\script.ps1 -SharepointURL $SharepointURL -tenantID $tenantID